+++ UnityProject/Assets/Scripts/Items/Weapons/GunTrigger.cs
@@ -1,56 +1,76 @@
-using System;
 using UnityEngine;
-
+using Mirror;
 
 namespace Weapons
 {
-	public class GunTrigger : MonoBehaviour
+	public class GunTrigger : NetworkBehaviour
 	{
 		[SerializeField]
-		private JobType setRestriction = default;
-		private Gun gun;
+		private JobType setRestriction;
 		[SerializeField]
-		private bool allowClumsy = default;
+		private bool allowClumsy;
 		[SerializeField]
-		private bool allowNonClumsy = default;
-		private readonly System.Random rnd = new System.Random();
+		private bool allowNonClumsy;
 
-		public void Awake()
-		{
-			gun = GetComponent<Gun>();
-		}
+		[HideInInspector, SyncVar(hook = nameof(SyncPredictionCanFire))]
+		public bool PredictionCanFire;
 
-		public void TriggerPull(GameObject shotBy, Vector2 target,
-			BodyPartType damageZone, bool isSuicideShot)
+		public int TriggerPull(GameObject shotBy)
 		{
 			JobType job = PlayerList.Instance.Get(shotBy).Job;
 
-			if (setRestriction == JobType.NULL)
+			if (job == setRestriction || (setRestriction == JobType.NULL &&
+			(job != JobType.CLOWN && allowNonClumsy || job == JobType.CLOWN && allowClumsy)))
 			{
-				if (job == JobType.CLOWN && !allowClumsy)
-				{
-					int chance = rnd.Next(0 ,2);
-					if (chance == 0)
-					{
-						gun.ServerShoot(shotBy , target, damageZone, true);
-						Chat.AddActionMsgToChat(
-							shotBy,
-							"You fumble up and shoot yourself!",
-							$"{shotBy.ExpensiveName()} fumbles up and shoots themself!");
-						return;
-					}
-				}
-				else if (job != JobType.CLOWN && !allowNonClumsy)
-				{
-					gun.ServerShoot(shotBy , target, damageZone, true);
-					Chat.AddActionMsgToChat(
-						shotBy,
-						"You somehow shoot yourself in the face! How the hell?!",
-						$"{shotBy.ExpensiveName()} somehow manages to shoot themself in the face!");
-					return;
-				}
+				// a normal shot
+				return 2;
 			}
-			gun.ServerShoot(shotBy , target, damageZone, isSuicideShot);
+			else if (setRestriction == JobType.NULL && (job == JobType.CLOWN && !allowClumsy))
+			{
+				//shooting a non-clusmy weapon as a clusmy person
+				return 3;
+			}
+			else if (setRestriction == JobType.NULL && (job != JobType.CLOWN && !allowNonClumsy))
+			{
+				return 1;
+			}
+			else
+			{
+				return 0;
+			}
 		}
+
+		public bool TriggerPullClient()
+		{
+			return PredictionCanFire;
+		}
+
+		//Serverside method to update syncvar
+		public void UpdatePredictionCanFire(GameObject holder)
+		{
+			JobType job = PlayerList.Instance.Get(holder).Job;
+			if (holder == null)
+			{
+				SyncPredictionCanFire(PredictionCanFire, false);
+			}
+			else if (job == setRestriction || setRestriction == JobType.NULL)
+			{
+				SyncPredictionCanFire(PredictionCanFire, true);
+			}
+		}
+
+		public void ClearPredictionCanFire()
+		{
+			SyncPredictionCanFire(PredictionCanFire, false);
+		}
+
+		/// <summary>
+		/// Syncs the prediction bool
+		/// </summary>
+		private void SyncPredictionCanFire(bool oldValue, bool newValue)
+		{
+			PredictionCanFire = newValue;
+		}
+
 	}
 }